requirements.txt
streamlit==1.32.2
pypdf==4.1.0
langchain==0.1.13
python-dotenv==1.0.1
chromadb>=0.4.22
langchain-google-genai==0.0.11